requirements.txt
openai
streamlit
pandas
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2 